feat(catalog): 🗂️ add subcategory structure and rename product files
- Created base files for Subcategory domain and infrastructure.
- Renamed generic product files (e.g., Repository to ProductRepository) to avoid God Classes and follow SRP.
- Updated namespaces and dependency injections for the new structure.
