--- CunDropShipping-Product/CunDropShipping-Products.http
@@ -1,6 +1,0 @@
-@CunDropShipping_HostAddress = http://localhost:5227
-
-GET {{CunDropShipping_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
