--- CunDropShipping-Product/appsettings.Development.json
@@ -1,11 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "DefaultConnection": "server=davidortega.dev;port=3306;database=davidort_cun_dropshipping;user=davidort_cun_dropshipping;password=davidort_cun_dropshipping;SslMode=none;"
-  }
-}
