+++ catalog-service.sln
@@ -1,0 +1,21 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "catalog-service", "catalog-service\catalog-service.csproj", "{90C24E8F-392A-40AD-AFD0-90D6FDA7B4B9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{21D9E7F0-9D53-495D-A4D1-8E921ACF3719}"
+	ProjectSection(SolutionItems) = preProject
+		compose.yaml = compose.yaml
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{90C24E8F-392A-40AD-AFD0-90D6FDA7B4B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{90C24E8F-392A-40AD-AFD0-90D6FDA7B4B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{90C24E8F-392A-40AD-AFD0-90D6FDA7B4B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{90C24E8F-392A-40AD-AFD0-90D6FDA7B4B9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
