+++ catalog-service/CunDropShipping-Products.http
@@ -1,0 +1,6 @@
+@Catalog_HostAddress = http://localhost:5227
+
+GET {{Catalog_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
