+++ catalog-service/Dockerfile
