+++ catalog-service/adapter/restful/v1/controller/Entity/AdapterSubCategoryEntity.cs
@@ -1,0 +1,8 @@
+﻿namespace Catalog.adapter.restful.v1.controller.Entity;
+
+public class AdapterSubCategoryEntity
+{
+    public int IdSubCategory { get; set; }
+    public string NameSubCategory { get; set; }
+    public int CategoryId { get; set; }
+}
