+++ catalog-service/adapter/restful/v1/controller/Mapper/ISubCategoryAdapterMapper.cs
@@ -1,0 +1,8 @@
+﻿using Catalog.adapter.restful.v1.controller.Entity;
+
+namespace Catalog.adapter.restful.v1.controller.Mapper;
+
+public interface ISubCategoryAdapterMapper
+{
+    AdapterSubCategoryEntity ToAdapterSubCategory(Domain)
+}
