+++ catalog-service/adapter/restful/v1/controller/Mapper/SubCategoryAdapterMapper.cs
@@ -1,0 +1,6 @@
+﻿namespace Catalog.adapter.restful.v1.controller.Mapper;
+
+public class SubCategoryAdapterMapper
+{
+    
+}
