+++ catalog-service/adapter/restful/v1/controller/SubCategoryController.cs
@@ -1,0 +1,6 @@
+﻿namespace CunDropShipping.adapter.restful.v1.controller;
+
+public class SubCategoryController
+{
+    
+}
