+++ catalog-service/application/Service/ISubCategoryService.cs
@@ -1,0 +1,6 @@
+﻿namespace Catalog.application.Service;
+
+public interface ISubCategoryService
+{
+    
+}
