+++ catalog-service/catalog-service.csproj
