+++ catalog-service/domain/Entity/DomainSubCategoryEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace Catalog.domain.Entity;
+
+public class DomainSubCategoryEntity
+{
+    
+}
