+++ catalog-service/domain/SubCategoryServiceImp.cs
@@ -1,0 +1,6 @@
+﻿namespace Catalog.domain;
+
+public class SubCategoryServiceImp
+{
+    
+}
