+++ catalog-service/infrastructure/DbContext/SubCategoryRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace Catalog.infrastructure.DbContext;
+
+public class SubCategoryRepository
+{
+    
+}
