+++ catalog-service/infrastructure/Mapper/ISubCategoryInfrastructureMapper.cs
@@ -1,0 +1,6 @@
+﻿namespace Catalog.infrastructure.Mapper;
+
+public interface ISubCategoryInfrastructureMapper
+{
+    
+}
