+++ catalog-service/infrastructure/Mapper/SubCategoryInfrastructureMapperImpl.cs
@@ -1,0 +1,6 @@
+﻿namespace Catalog.infrastructure.Mapper;
+
+public class SubCategoryInfrastructureMapperImpl
+{
+    
+}
